--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/old/pipeline/SleeveDetectionRight.java
@@ -1,144 +1,0 @@
-package org.firstinspires.ftc.teamcode.old.pipeline;
-
-import org.opencv.core.Core;
-import org.opencv.core.Mat;
-import org.opencv.core.Point;
-import org.opencv.core.Rect;
-import org.opencv.core.Scalar;
-import org.opencv.core.Size;
-import org.opencv.imgproc.Imgproc;
-import org.openftc.easyopencv.OpenCvPipeline;
-
-public class SleeveDetectionRight extends OpenCvPipeline {
-    /*
-    YELLOW  = Parking Left
-    CYAN    = Parking Middle
-    MAGENTA = Parking Right
-     */
-
-    public enum ParkingPosition {
-        LEFT,
-        CENTER,
-        RIGHT
-    }
-
-    // TOPLEFT anchor point for the bounding box
-    private static final Point SLEEVE_TOPLEFT_ANCHOR_POINT = new Point(125, 175);
-
-    // Width and height for the bounding box
-    public static int REGION_WIDTH = 25;
-    public static int REGION_HEIGHT = 15;
-
-    // Lower and upper boundaries for colors
-    private static final Scalar
-//            lower_yellow_bounds = new Scalar(150, 150, 0, 255),
-//            upper_yellow_bounds = new Scalar(255, 255, 150, 255),
-//            lower_cyan_bounds = new Scalar(0, 120, 120, 255),
-//            upper_cyan_bounds = new Scalar(150, 255, 255, 255),
-//            lower_magenta_bounds = new Scalar(120, 0, 120, 255),
-//            upper_magenta_bounds = new Scalar(255, 170, 255, 255);
-
-            lower_yellow_bounds = new Scalar(120, 120, 0, 255),
-            upper_yellow_bounds = new Scalar(255, 255, 70, 255),
-            lower_cyan_bounds = new Scalar(2, 82, 176, 255),
-            upper_cyan_bounds = new Scalar(48, 255, 255, 255),
-            lower_magenta_bounds = new Scalar(125, 3, 176, 255),
-            upper_magenta_bounds = new Scalar(255, 43, 255, 255);
-
-    // Color definitions
-    private final Scalar
-            YELLOW = new Scalar(255, 255, 0),
-            CYAN = new Scalar(0, 255, 255),
-            MAGENTA = new Scalar(255, 0, 255);
-
-    // Percent and mat definitions
-    private double yelPercent, cyaPercent, magPercent;
-    private final Mat yelMat = new Mat();
-    private final Mat cyaMat = new Mat();
-    private final Mat magMat = new Mat();
-    private Mat blurredMat = new Mat();
-    private Mat kernel = new Mat();
-
-    // Anchor point definitions
-    Point sleeve_pointA = new Point(
-            SLEEVE_TOPLEFT_ANCHOR_POINT.x,
-            SLEEVE_TOPLEFT_ANCHOR_POINT.y);
-    Point sleeve_pointB = new Point(
-            SLEEVE_TOPLEFT_ANCHOR_POINT.x + REGION_WIDTH,
-            SLEEVE_TOPLEFT_ANCHOR_POINT.y + REGION_HEIGHT);
-
-    // Running variable storing the parking position
-    private volatile ParkingPosition position = ParkingPosition.LEFT;
-
-    @Override
-    public Mat processFrame(Mat input) {
-        // Noise reduction
-
-        Imgproc.blur(input, blurredMat, new Size(5, 5));
-        blurredMat = blurredMat.submat(new Rect(sleeve_pointA, sleeve_pointB));
-
-        // Apply Morphology
-        kernel = Imgproc.getStructuringElement(Imgproc.MORPH_RECT, new Size(3, 3));
-        Imgproc.morphologyEx(blurredMat, blurredMat, Imgproc.MORPH_CLOSE, kernel);
-        Imgproc.cvtColor(input, blurredMat, Imgproc.COLOR_RGB2HSV);
-        blurredMat.convertTo(blurredMat, -1, 1, 10);
-
-        // Gets channels from given source mat
-        Core.inRange(blurredMat, lower_yellow_bounds, upper_yellow_bounds, yelMat);
-        Core.inRange(blurredMat, lower_cyan_bounds, upper_cyan_bounds, cyaMat);
-        Core.inRange(blurredMat, lower_magenta_bounds, upper_magenta_bounds, magMat);
-
-        // Gets color specific values
-        yelPercent = Core.countNonZero(yelMat);
-        cyaPercent = Core.countNonZero(cyaMat);
-        magPercent = Core.countNonZero(magMat);
-
-        // Calculates the highest amount of pixels being covered on each side
-        double maxPercent = Math.max(yelPercent, Math.max(cyaPercent, magPercent));
-
-        // Checks all percentages, will highlight bounding box in camera preview
-        // based on what color is being detected
-        if (maxPercent == yelPercent) {
-            position = ParkingPosition.LEFT;
-            Imgproc.rectangle(
-                    input,
-                    sleeve_pointA,
-                    sleeve_pointB,
-                    YELLOW,
-                    2
-            );
-        } else if (maxPercent == cyaPercent) {
-            position = ParkingPosition.CENTER;
-            Imgproc.rectangle(
-                    input,
-                    sleeve_pointA,
-                    sleeve_pointB,
-                    CYAN,
-                    2
-            );
-        } else if (maxPercent == magPercent) {
-            position = ParkingPosition.RIGHT;
-            Imgproc.rectangle(
-                    input,
-                    sleeve_pointA,
-                    sleeve_pointB,
-                    MAGENTA,
-                    2
-            );
-        }
-
-        // Memory cleanup
-        blurredMat.release();
-        yelMat.release();
-        cyaMat.release();
-        magMat.release();
-        kernel.release();
-
-        return input;
-    }
-
-    // Returns an enum being the current position where the robot will park
-    public ParkingPosition getPosition() {
-        return position;
-    }
-}
