--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/TurnTest.java
@@ -1,27 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmode;
-
-import com.acmerobotics.dashboard.config.Config;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
-
-/*
- * This is a simple routine to test turning capabilities.
- */
-@Config
-@Autonomous(group = "drive")
-public class TurnTest extends LinearOpMode {
-    public static double ANGLE = 90; // deg
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-
-        waitForStart();
-
-        if (isStopRequested()) return;
-
-        drive.turn(Math.toRadians(ANGLE));
-    }
-}
